Sandbox Setup/requirements.txt
cloudshell-orch-core>=2.1.0.0,<2.2.0.0
cloudshell-orch-ixia>=1.0,<1.1